gossip: move node to gossip folder
