+++ gossip/node.h
